rev.3

+++ ewm-service/src/main/java/ru/practicum/ewm/controller/priv/PrivateRequestController.java
@@ -17,23 +17,23 @@ public class PrivateRequestController {
 
     // Получение информации о заявках текущего пользователя на участие в чужих событиях
     @GetMapping
-    public List<ParticipationRequestDto> getRequests(@PathVariable int userId) {
+    public List<ParticipationRequestDto> getRequests(@PathVariable long userId) {
         log.trace("Получен GET-запрос по заявкам пользователя ID {}.", userId);
         return service.getRequests(userId);
     }
 
     // Добавление запроса от текущего пользователя на участие в событии
     @PostMapping
-    public ParticipationRequestDto addRequest(@PathVariable int userId,
-                                              @RequestParam(name = "eventId") int eventId) {
+    public ParticipationRequestDto addRequest(@PathVariable long userId,
+                                              @RequestParam(name = "eventId") long eventId) {
         log.trace("Получен POST-запрос на добавление заявки на событие ID {} от пользователя ID {}.", eventId, userId);
         return service.addRequest(userId, eventId);
     }
 
     // Отмена своего запроса на участие в событии
     @PatchMapping("/{requestId}/cancel")
-    public ParticipationRequestDto cancelRequest(@PathVariable int userId,
-                                                 @PathVariable int requestId) {
+    public ParticipationRequestDto cancelRequest(@PathVariable long userId,
+                                                 @PathVariable long requestId) {
         log.trace("Получен PATCH-запрос на отмену заявки ID {} пользователя ID {}.", requestId, userId);
         return service.cancelRequest(userId, requestId);
     }
